requirements.txt
cython==0.29.36
kivy==2.3.0
kivymd==1.2.0
openai
requests
pillow
cryptography
pyjnius
buildozer
opencv-python==4.8.1.78
numpy==1.24.3